# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "领域知识个性化生成与多智能体协同决策系统",
    "description": "领域知识个性化生成与多智能体协同决策系统后端API\n\n## 认证方式\n- 除 `/auth/login`、`/auth/register`、健康检查外，所有接口需在请求头携带 `Authorization: Bearer <access_token>`\n- Token 通过 `POST /api/v1/auth/login` 获取，有效期由 `JWT_EXPIRE_MINUTES` 控制\n- Token 过期后通过 `POST /api/v1/auth/refresh` 刷新\n\n## 统一响应格式\n所有接口返回 `{code, message, data, timestamp}` 结构，`code=200` 表示成功",
    "contact": {
      "name": "知识系统团队",
      "url": "https://github.com/knowledge-system"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "当前环境（相对路径）"
    },
    {
      "url": "http://localhost:8000/api/v1",
      "description": "开发环境"
    },
    {
      "url": "https://staging.example.com/api/v1",
      "description": "预发布环境"
    },
    {
      "url": "https://api.example.com/api/v1",
      "description": "生产环境"
    }
  ],
  "paths": {
...
    "/api/v1/resources/generate": {
      "post": {
        "tags": [
          "个性化资源生成"
        ],
        "summary": "生成三类个性化学习资源（Celery异步）",
        "description": "一次性生成三类学习资源：定制实操指南、分阶训练测试题、专属知识讲义\n\n- 传入学情画像ID，自动调度多Agent协同\n- 复用诊断+检索阶段，避免重复执行\n- 自动计算资源匹配度\n- 优先使用 Celery 异步队列，不可用时降级为 BackgroundTasks",
        "operationId": "generate_resources_api_v1_resources_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateResourcesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/resources/generate/batch": {
      "post": {
        "tags": [
          "个性化资源生成"
        ],
        "summary": "批量生成资源（Celery异步）",
        "description": "批量生成资源（多个学习者）\n\n请求体:\n{\n    \"learner_ids\": [1, 2, 3],\n    \"target_topic\": \"Python机器学习\",\n    \"industry\": \"人工智能训练\"\n}",
        "operationId": "batch_generate_resources_api_v1_resources_generate_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "title": "Request"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/tasks/{task_id}/status": {
      "get": {
        "tags": [
          "个性化资源生成"
        ],
        "summary": "查询 Celery 任务进度",
        "description": "查询异步任务进度\n\n- 返回当前阶段、进度百分比、状态信息\n- 支持 resource_generation / batch_generation 等任务类型",
        "operationId": "get_task_status_api_v1_tasks__task_id__status_get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/resources/generate/sync": {
      "post": {
        "tags": [
          "个性化资源生成"
        ],
        "summary": "同步生成三类资源",
        "description": "同步生成三类学习资源（演示用）\n\n- 传入学情画像ID，自动调度多Agent协同\n- 返回完整生成结果",
        "operationId": "generate_resources_sync_api_v1_resources_generate_sync_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateResourcesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/resources": {
      "get": {
        "tags": [
          "个性化资源生成"
        ],
        "summary": "获取资源列表",
        "description": "获取资源列表（支持分页、筛选）\n\n- 支持按学习者、类型、难度、状态筛选\n- 返回资源匹配度等关键指标",
        "operationId": "get_resource_list_api_v1_resources_get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "learner_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "学习者ID",
              "title": "Learner Id"
            },
            "description": "学习者ID"
          },
          {
            "name": "resource_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "资源类型",
              "title": "Resource Type"
            },
            "description": "资源类型"
          },
          {
            "name": "difficulty_level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "难度等级",
              "title": "Difficulty Level"
            },
            "description": "难度等级"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "状态",
              "title": "Status"
            },
            "description": "状态"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "页码",
              "default": 1,
              "title": "Page"
            },
            "description": "页码"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "每页大小",
              "default": 20,
              "title": "Page Size"
            },
            "description": "每页大小"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/resources/{resource_id}": {
      "get": {
        "tags": [
          "个性化资源生成"
        ],
        "summary": "获取资源详情",
        "description": "获取资源详情（含完整内容）\n\n- 返回资源完整信息，包括内容、匹配度、来源切片等",
        "operationId": "get_resource_detail_api_v1_resources__resource_id__get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "resource_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Resource Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/resources/{resource_id}/export": {
      "get": {
        "tags": [
          "个性化资源生成"
        ],
        "summary": "导出资源",
        "description": "导出资源文件\n\n- 支持txt和md两种格式\n- 返回纯文本内容，可直接下载",
        "operationId": "export_resource_api_v1_resources__resource_id__export_get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "resource_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Resource Id"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "导出格式: txt/md",
              "default": "txt",
              "title": "Format"
            },
            "description": "导出格式: txt/md"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BaseResponse": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "状态码",
            "default": 200
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "提示消息",
            "default": "操作成功"
          },
          "data": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "业务数据"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp",
            "description": "响应时间戳"
          }
        },
        "type": "object",
        "title": "BaseResponse",
        "description": "统一返回结果基类",
        "example": {
          "code": 200,
          "message": "操作成功",
          "timestamp": "2024-03-15 14:30:00"
        }
      },
...
      "GenerateResourcesRequest": {
        "properties": {
          "learner_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Learner Id",
            "description": "学习者ID"
          },
          "target_topic": {
            "type": "string",
            "minLength": 1,
            "title": "Target Topic",
            "description": "目标主题"
          },
          "industry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Industry",
            "description": "行业领域"
          }
        },
        "type": "object",
        "required": [
          "learner_id",
          "target_topic"
        ],
        "title": "GenerateResourcesRequest",
        "description": "生成资源请求"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  },
...
}